remove non-test chai dependency
