+++ src/utils.ts
@@ -1,4 +1,3 @@
-import { assert } from 'chai';
 import { AddressInfo, Server } from 'net';
 
 export const ZERO_ADDRESS = '0x0000000000000000000000000000000000000000';
@@ -18,34 +17,6 @@ export function isNumbers(array: number[] | any[]): array is number[] {
 
 export function isStrings(array: string[] | any[]): array is string[] {
     return array.length > 0 && typeof array[0] === 'string';
-}
-
-export function assertDeepEqual(a: any, b: any, ignore: string[], message?: string) {
-    if (Array.isArray(a) && Array.isArray(b)) {
-        const a1 = a.map(x => {
-            const x1 = { ...x };
-            ignore.forEach(f => {
-                x1[f] = undefined;
-            });
-            return { ...x1 };
-        });
-        const b1 = b.map(x => {
-            const x1 = { ...x };
-            ignore.forEach(f => {
-                x1[f] = undefined;
-            });
-            return { ...x1 };
-        });
-
-        assert.deepEqual(a1, b1, message);
-    } else {
-        const a1 = { ...a };
-        const b1 = { ...b };
-        ignore.forEach(f => {
-            (a1[f] = undefined), (b1[f] = undefined);
-        });
-        assert.deepEqual(a1, b1, message);
-    }
 }
 
 export const sleep = (ms: number) => {
